refactor: extract GettingStartedPage logic into useGettingStartedProgress hook (F5)
Move the progress state, localStorage load/save, keyboard-nav effects, and
module/chapter navigation into a page-local hook
(components/getting-started/useGettingStartedProgress.js) with its
STORAGE_KEY/loadProgress/saveProgress helpers. GettingStartedPage becomes a
render shell (626 -> 478 lines) keeping its multi-view JSX and getMasteryLevel.

Behavior-preserving: hook body moved verbatim (same hook order/effect deps),
no JSX guards, returning exactly the state and handlers the JSX consumes.
Verified by the existing test, full suite (83), lint, and build. This
completes page-local data/logic hooks across all the large pages.
